+++ src/main/java/konkuk/thip/recentSearch/application/service/manager/RecentSearchCreateManager.java
@@ -1,6 +1,6 @@
 package konkuk.thip.recentSearch.application.service.manager;
 
-import konkuk.thip.common.annotation.HelperService;
+import konkuk.thip.common.annotation.application.HelperService;
 import konkuk.thip.recentSearch.adapter.out.jpa.RecentSearchType;
 import konkuk.thip.recentSearch.application.port.out.RecentSearchCommandPort;
 import konkuk.thip.recentSearch.application.port.out.RecentSearchQueryPort;
