+++ src/main/java/konkuk/thip/room/application/service/validator/RoomParticipantValidator.java
@@ -1,6 +1,6 @@
 package konkuk.thip.room.application.service.validator;
 
-import konkuk.thip.common.annotation.HelperService;
+import konkuk.thip.common.annotation.application.HelperService;
 import konkuk.thip.common.exception.InvalidStateException;
 import konkuk.thip.room.application.port.out.RoomParticipantQueryPort;
 import lombok.RequiredArgsConstructor;
